externalized submit function to declutter AddArticleForm component
